add libdrm

+++ tools/generate_source.sh
@@ -2,10 +2,19 @@
 
 . version.sh
 
-git clone ${git_repo} ${package}
-if [ -f ./${package}/.git/config ] ; then
-	cd ./${package}/
-	git archive --format=tar --prefix=${package}_${version}/ HEAD | xz > ../${package}_${version}.orig.tar.xz
-	cd ../
-	rm -rf ./${package}/
+if [ ! "x${git_repo}" = "x" ] ; then
+	git clone ${git_repo} ${package}
+	if [ -f ./${package}/.git/config ] ; then
+		cd ./${package}/
+		git archive --format=tar --prefix=${package}_${version}/ HEAD | xz > ../${package}_${version}.orig.tar.xz
+		cd ../
+		rm -rf ./${package}/
+	fi
+fi
+
+if [ ! "x${deb_source}" = "x" ] ; then
+	wget -c ${deb_source}
+	if [ ! "x${deb_patch}" = "x" ] ; then
+		wget -c ${deb_patch}
+	fi
 fi
